refactor(dicts): rename urlDIct -> URL_DICT
